Use fedorapeople for getting fedora image
It looks the official Fedora mirror is unstable right now, so we
uploaded the image to fedorapeople. This is a temprorary solution.
Once the official mirror become stable, we switch back to it.

Note: alt.fedoraproject.org seems to have a newer image and
the official mirrors (download.fedoraproject.org) do not seem to have
gotten the new image yet. So when download.fedoraproject.org
redirects (302) to a mirror the image does not exist there. We can
pull image from alt directly, but we opted for using fedorapeople as
we don't want to introduce load on alt.

Co-Authored-By: Hongbin Lu <hongbin.lu@huawei.com>

Closes-Bug: #1567124
Change-Id: I441898d846d1768b9723b293196b5339271a377e