Update Kube version for latest image
New image on 4/5 has Kube version 1.2.0

For Kubernetes bay, the scheduler, controller and proxy
are run from container, not from Kube binaries in the
image.  The version for these containers are currently
hardcoded in the Heat template as 1.0.6, so update to
the new version 1.2.0.  Otherwise, the scheduler and
kubelet would not work properly with the mixed versions.

There is a patch under review to expose the Kube version
as a label:
https://review.openstack.org/#/c/295655/

Implements: blueprint atomic23
Change-Id: I46be4038b62be202ba670749e0a5c6663eeb12d3