Add Storage section in user guide
Describe how to use Cinder block with COE's

This section covers both ephemeral and persistent storage for
containers.  It describes how Cinder block is used by Kubernetes,
Swarm, Mesos, and how Magnum is providing support along with
the options for users to manage the storage.  Examples are
included.

The doc container-volume-integration.rst is updated to the
current status and incorporated into the user guide.

Partially implements: blueprint user-guide
Change-Id: Iecd1a58e7bc04dda905e4c43d709f9eec7f68b97