Update Image section in user guide
New details for building Fedora Atomic image

There is a new automated build process for the Fedora Atomic image
for Kubernetes and Swarm, based on diskimagebuilder.  Remove the
document for manual build since it is outdated, and update the
user guide accordingly, pointing to the new instruction using
diskimagebuilder.

Partially implements: blueprint user-guide
Change-Id: Ie8bc70989f7322a6184e2ab1cf9653ad79f92f24