Register k8s node but make it unschedulable
In before, we configure k8s master node by using
"--register-node=false". This seems to work but will produce verbose
error message in the log [1]. Starting from v1.2, there is a new
option "--register-schedulable". We can use the combination of both
options to register the master node but make it unschedulable. This
can suppress the verbose error message.

[1] https://github.com/kubernetes/kubernetes/issues/15108

Closes-Bug: #1584198
Change-Id: I72cc360e08b647bf239b62af6a7fd87ca14c604e