Put fault info of bay resources into bay-show outputs
In order to debug a failing bay creation, users need to know about
interaction with heat, heat architecture and how to debug it.
It would be much better if magnum could propagate some of the error
description into the bay-show status.

This patch proposes a way to do this:
If a bay was in error status('CREATE_FAILED', 'UPDATE_FAILED',
'DELETE_FAILED'), we collect fault info from heat resources of the
bay and put the fault info into 'bay-show' command outputs.

Change-Id: I458f1619243bf44f5f0e7dd2a5d6cd10ed2d0fde
Partially-Implements: blueprint explain-errors