Fix bug for write-kube-os-config.sh
In file write-kube-os-config.sh, the bug is in line below
AUTH_URL=$(echo "$AUTH_URL" | tr -s "v3" "v2")
this command is used to change v3 to v2,but it will change any "3" to
"2". For example:
AUTH_URL=10.43.166.108:5000/v3
it will be changed to 10.42.166.108:5000/v2, where 43 is also changed
to 42.

Change-Id: Ia329b983ba59856783f2873722c9ade418d49cea
Closes-Bug: #1599703