Fix copying logs from nodes
`_get_nodes_from_stack` will be never called because `_get_nodes_from_bay`
always return un-empty list if master_addresses and node_addresses are
None. So we can't get logs unless bay is succeeded to create.
This patch fixes this issue, so we can get node logs if bay is not create
completed.

Change-Id: I6e518359d08cf06e95adc0b324c934c8f3e0c27d
Closes-Bug: #1605092