Openvswitch image build
Dockerfile to build Openvswitch image

The Kuryr driver for Swarm bay requires Openvswitch and the Neutron
L2 agent running on the nodes.  Since the Fedora Atomic does not
have these packages installed and it is preferrable to not use a
custom-built image,  we need to run these packages in a container.
This Dockerfile is used to build the Docker image hosted on
Docker Hub:

openstackmagnum/fedora23-neutron-ovs:testing

Partially implements: blueprint kuryr-swarm-integration
Change-Id: I5132ccc2bc5660aa9688430bbc51f6759f6b7a81