Simplify test_create_list_sign_delete_clusters() tempest test
The test needs a valid CSR which was created in the devstack post_test_hook
in a static way (always the same CSR).
This commit simplifies the process and just adds the CSR to the test itself
so that the CSR string can be used to check the CSR. That way there is no need
to provide a file with the CSR content and point the tempest configuration
via the "csr_location" parameter to that file.
So the "csr_location" parameter is now deprecated for removal.
Also remove the unused "invalid.csr" which was also created in the
post_test_hook.

Change-Id: I907130576cfb59267643b4639681331acfcfa07f