Create sysconfig mount for kubernetes controller mgr
Since the kubernetes controller manager has been moved to a separate
pod, that pod needs to have access to /etc/sysconfig/kube_openstack_config
in order for the kubernetes load balancer feature to work. The
kube_openstack_config file contains the Kubernetes load balancer
configuration as well as the OpenStack credentials for the Neutron
client.

Note that as an alternative to this proposed fix, we could instead
change the Magnum developer's documentation for using the kubernetes
external load balancer feature to suggest having the cluster administrator
set up the mount point dynamically.

Change-Id: I407f5c7ba2069906f7bdbc4abb1c7618fa1e08e8
Partial-Bug: #1524025