Update name of fedora-atomic image
Based on past experience, some failures are caused by using the
stable branch with latest image. That is because the latest image
are supposed to use with master branch.

Let's make this clear by renaming the image from
"fedora-atomic-latest" to "fedora-atomic-newton" in this branch.
In fedorapeople, a soft link was created from
"fedora-atomic-newton.qcow2" to "fedora-23-atomic-20160405.qcow2"

Change-Id: I10e5594b5c5a3ffb9483e7b49bf4325a3c066b1c