Update name of ubuntu-mesos image
Based on past experience, some failures are caused by using the
stable branch with latest image. That is because the latest image
are supposed to use with master branch.

Let's make this clear by renaming the image from
"ubuntu-14.04.3-mesos-0.25.0" to "ubuntu-mesos-newton" in this branch
In fedorapeople, a soft link was created from
"ubuntu-mesos-newton.qcow2" to "ubuntu-14.04.3-mesos-0.25.0.qcow2"

Change-Id: Ib80e29312ebdb8c420875c9fe4bd1d4f2c28453e