Make cinder volume optional
The purpose of this patch is primarily to unblock the
stable/newton gate and secondarily fix the performance issues
faced with large clusters, where Magnum creates a volume per
node.

In the swarm_atomic and k8s_atomic drivers container images are
stored in a dedicated cinder volume per cluster node. It is
proven that this architecture can be a scalability bottleneck.

Make the use of cinder volumes for container images and opt-in
option.  If docker-volume-size is not specified no cinder
volumes will be created. Before, if docker-volume-size wasn't
specified the default value was 25.

To use cinder volumes for container storage the user will
interact with magnum as before, (meaning the valid values are
integers starting from 1).

Backport: I3394c62a43bbf950b7cf0b86a71b1d9b0481d68f
Conflicts:
* magnum/drivers/common/swarm_fedora_template_def.py
  Edit magnum/drivers/swarm_fedora_atomic_v1/template_def.py
  instead.
* magnum/tests/unit/conductor/handlers/test_k8s_cluster_conductor.py
  Remove invalid unit test for docker_volume_size.
  Fix unit test which references the driver class

Additionally, remove the use of cinder volumes in functioanal tests.
2nd Backport: Ia3b14603c5fc516b00c862c8b9257e0fd23d4b9e

Remove service from manager class for tempest.
3rd Backport: I67f79efd2049c05d36ea56691b664417ed358fd8

Closes-Bug: #1638006
Related-Bug: #1422831

Change-Id: I219f02dc1861bd4b6b9c59ecc6af448d09004f18