func-test-docs: Use iniget and set concurrecy 1
In devstack, we have host in Default and in api. To get the HOST_IP
we must look in the api section. Use iniget to take the IP instead of
awk, making more concise where the IP comes from.

Recommend to use --concurrency 1 when running tests to avoid creating
multiple clusters.

Change-Id: I9b38bc88e4157e6692d7cffaed538c4c0a769d03