Disable horizon, ceilomter and swift in gate hook
If the services are disabled in the devstack/settings file, it disables
them for everyone who uses the magnum devstack plugin. Some people (like
shade) use both magnum and swift to test things. Putting this in the
gate hook means it's used in all of magnum's jobs, but not in other
people's jobs.

Change-Id: Icd82a4ab68533f39f967575c2125b0f61c70e0a6