Magnum Development Policies
This patch introduces a set of responsibilities, guidelines, and
policies for our conduct. It outlines our team's philosophy for
development in this stage of our maturity cycle.

It extends http://http://docs.openstack.org/project-team-guide with
care taken not to conflict with it.

This may be used for discussion at our team discussion at our upcoming
PTG session on Ocata Retrospective / Team Dynamics.

Change-Id: I4878672c99fec0e5ea8c4beb0001cb9dfb4a30f5