Don't enforce microversion for stats API
Due to bug in microversion logic, adding any new microversioned
API would mandate OpenStack-API-Version header in the request.
This change is to avoid microversion enforcement until the
bug in microversioning is fixed.
Please note that, this issue is only seen when no version
header is sent to the controller the first time and also
this bug can not be reproduced by magnum CLI since it sends
the latest version header by default.

Change-Id: I9a576e048846e930c62fb55584b125585137fcfd
Partial-Bug: #1659431