Fix database grant instructions in install guide
On at least mariadb, there is a default grant that restricts connections
from localhost and supercedes other permissions. In order for the db
sync command to work later on, we need to specifically enable
connections from localhost for the 'magnum' user. Moreover, it is
redundant to grant permissions for hostname 'controller' and the
wildcard host '%'. This patch fixes the error and makes magnum's
database configuration instructions consistent with the main install
guide

Change-Id: I21f2c9a2f3978ae6011b4b7b49c8edd826781284