Use eventlet executor in rpc_service
The default executor of o.m is 'blocking', but after change [1],
o.m with blocking is broken. 'eventlet' and 'non-blocking' work
fine and most services use 'eventlet', so this is why it was
noticed by magnum which used the default one.

Use eventlet to not be affected by bug #1694728 , but more
importantly, the oslo team suggests to not use blocking which
has 0% test coverage and it is going to be deprecated and unset
from the default configuration.

[1] https://review.openstack.org/#/c/463673/

Change-Id: I47da73787456c97f7d84fd4440404b551ff62528
Closes-Bug: #1694728
(cherry picked from commit bd69b3fff6b50048813360627fb02044cf1bcb58)