Fix cluster inheritence of docker_volume_size
Following [1], cluster objects have a docker_volume_size attribute.
If not explicitly set on creation, the value of this attribute should
default to the value of the same field on the cluster's template.

When not provided, the API Cluster object's docker_volume_size field
takes the value wsme.Unset, rather than None.

[1]
https://github.com/openstack/magnum/commit/79039bb4198b08500f5d5738b464d187f494a73a

Change-Id: Iad5231643c01ec9698c932b59806781034390795
Depends-On: I510ff10d708a237be0cb88d3ff7094d7c81f5875
Closes-Bug: #1702314