Don't poll heat if no stack exists
Cluster objects are created asynchronously from their underlying
heat stacks, meaning that the periodic update can sometimes end up
trying to poll a cluster's heat stack before the stack has been created.

This change checks whether the stack_id is None and skips polling heat
if so. This has the side effect of resolving bug 1682058, since we don't
try to use a trust and trustee that do not exist.

Change-Id: I73f039659250f1d5b69b23141835c4602c8e019a
Closes-Bug: #1682058