Add default for [cinder]default_docker_volume_type
When creating a kubernetes cluster on baremetal & fedora, if the cluster
template does not have a docker_volume_type defined, the following error
is seen:

InvalidParameterValue: ERROR: The Parameter (docker_volume_type) was not
provided.

Cinder isn't mandatory, and neither is the docker_volume_type cluster
template parameter, so it shouldn't need to be set.

This change adds a default value of an empty string for the option
[cinder]default_docker_volume_type, which allows the cluster to be
created.

Change-Id: I4416e2826e4a14a11b93d55d342e4de9b3dc12d7
Closes-Bug: #1702075