k8s: Fix apiserver configuration
kube-apiserver configuration was modified using sed and / as a
delimiter. However, the arguments were containing / so the sed command
was breaking.

Furthermore, the KUBE_API_ARGS was added without quotes which was
working when using the file in  a systemd unit, it does not work when
sourcing with . or source.

Implements: blueprint run-kube-as-container

Change-Id: I49714f51159ea1bf4e9918bbec82749443cc1751