Remove /etc/ssl/certs in the controller manager pod
For system such as Fedora Atomic, the CA bundle files which are
contained in /etc/ssl/certs are symbolic links to /etc/pki. When
configuring the controller manager to use an SSL endpoint, it will
raise an error as it is unable to authenticate the SSL endpoint.

This patch removes the host mount at /etc/ssl/certs.  The Hyperkube
images already ship a collection of CAs which are likely good for
all needs.

Closes-Bug: #1708452

Change-Id: Ife2b60d1968482a8c3ab9b44abbe401c6230881c