Now user can update label values in cluster-template
In Magnum Labels are stored in the form of dictionary. previously we
are passing string value directly to store the value of label. Now we
are parsing it and storing it in the form of dictionary.

Change-Id: I4d64da78dc4ed4d5599533b54861b65bce609c28
Closes-Bug: #1638863
(cherry picked from commit 21c87f35a0910963ac525dca829e98fe9f9a73f8)