Start RPC service before waiting
Stoping magnum-cond without having invoke start(),
results in "WARNING oslo_messaging.server Possible
hang: stop is waiting for start to complete".

A magnum instance with 16 workers was taking 1m to stop
with this change it takes 1 to 10 seconds. This change
doesn't break the fix in [1].

[1] If9e13effc14fd35e646d02bb4f568e79786aa958

Related-Bug: #1702349
Related issue in sahara:
Related-Bug: #1546119

Change-Id: Ied7ab43398d4e499514fa0bd5dba64971d1956bf