Add missing RBAC config for Prometheus
Currently the prometheus service is broken because the RBAC config is
missing. The patch just adds it to make it work again.

And the namespace of prometheus and grafana are also changed in this
patch to get a better isolation.

Closes-Bug: #1755772

Change-Id: I05affd9091ad13d70cc074b45d204162f2f4b21e
(cherry picked from commit d07fb7f3cc0c16ad597e930c37b042c26ac78049)