k8s_fedora: Add cloud_provider_enabled label
Add 'cloud_provider_enabled' label for the k8s_fedora_atomic
driver. Defaults to true. For specific kubernetes versions if
'cinder' is selected as a 'volume_driver', it is implied that
the cloud provider will be enabled since they are combined.

The motivation for this change is that in environments with
high load to the OpenStack APIs, users might want to disable
the cloud provider.

story: 1775358
task: 20253

Change-Id: I2920f699654af1f4ba45644ab60a04a3f70918fe
(cherry picked from commit 974399a912b02ebe1587ad932a405ca8b44dc947)