[k8s] Set order in kubemaster software deployments
When we create a cluster and pass the ca.key in a software deployment we
must ensure that the apiserver will start before calico, dashboard etc
which require the api to return ok. [0]

The heat agent process the deployment serially, so if coredns arrives
first in the agent, it will wait forever for the coredns script to
complete.

Putting the cert_manager_api first solves the issue.

[0] curl http://127.0.0.1:8080/healthz

Change-Id: I031ab34141045dde171bcf6206e227fa7eb5885d
story: 2003434
task: 24630