functional: use default admission_control_list values
The functional jobs are currently deploying with an empty list of
admission control list which means that the service token controller
does not go up resulting in the cluster failing to go up.

This patch drops that so it uses the default values of the admission
controllers in order to get the cluster to go up cleanly.

Change-Id: I0fdd65a9859f34e202016c37620f553623ef8a3e