Fix openstack-cloud-controller-manager restarts
Openstack-cloud-controller-manager restarts several times during
cluster creation.

This happens because cloud-controller-manager starts running before
needed secrets exist in kubernetes. Cloud-controller-manager lists secrets
and if the secrets exists it uses it and moves on, but if the secret
doesn't exist it starts a watch until it does. As this is not allowed the
pod fails.

This is triggered by Issue 
https://github.com/kubernetes/cloud-provider-openstack/issues/545

Story: 2005270

Change-Id: If8f34dc45b3b8a76e3d561ed41b4d0a783ceecb5
Signed-off-by: Diogo Guerra <dy090.guerra@gmail.com>
(cherry picked from commit 21acb8dc9afbc3c1b2a786c4041213fa7e8761dc)