Specify internal network to improve stability in a multi-NIC scenario.
When there is more than one NIC attached to an instance, openstack cloud
provider returns a random InternalIP back to the host resulting in instability
with API server which only talks to a default interface.

This patch incorporates the changes made in
https://github.com/kubernetes/cloud-provider-openstack/pull/444 which enables
OpenStack Cloud Controller Manager to respect the `internal-network-name` in
cloud-config file which ensures that InternalIP remains stable.

Story: 2005333
Task: 30271

Change-Id: I9e3ad459dd05753b53cb4ce75ee3aed649fef196